requirements.txt
discord.py>=1.3
git+https://github.com/Rapptz/discord-ext-menus >= 1.0